% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{ABS_METRICS}
\alias{ABS_METRICS}
\title{Metrics that should be absolute-valued before rank normalisation}
\format{
An object of class \code{character} of length 6.
}
\usage{
ABS_METRICS
}
\description{
Metrics that should be absolute-valued before rank normalisation
}
\keyword{internal}
