% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{CONFIDENCE_THRESHOLDS}
\alias{CONFIDENCE_THRESHOLDS}
\title{Confidence classification thresholds (based on adjusted p-values)}
\format{
An object of class \code{list} of length 3.
}
\usage{
CONFIDENCE_THRESHOLDS
}
\description{
Confidence classification thresholds (based on adjusted p-values)
}
\keyword{internal}
