% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{DEFAULT_PARAMS}
\alias{DEFAULT_PARAMS}
\title{Default analysis parameters}
\format{
An object of class \code{list} of length 13.
}
\usage{
DEFAULT_PARAMS
}
\description{
Default analysis parameters
}
\keyword{internal}
