% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{DEFAULT_WEIGHTS}
\alias{DEFAULT_WEIGHTS}
\title{Default metric weights for score integration}
\format{
An object of class \code{numeric} of length 14.
}
\usage{
DEFAULT_WEIGHTS
}
\description{
Default metric weights for score integration
}
\keyword{internal}
