% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_celltype.R
\name{.assign_microenv_bins}
\alias{.assign_microenv_bins}
\title{Assign cells to micro-environment bins using k-means on PCA embedding}
\usage{
.assign_microenv_bins(embed, n_bins)
}
\arguments{
\item{embed}{Numeric matrix (n_cells x n_dims) of PCA (or other) embedding.}

\item{n_bins}{Integer; number of micro-environment bins.}
}
\value{
Integer vector of bin assignments (1..n_bins).
}
\description{
Cells are clustered into spatial micro-environment bins so that each bin
represents a local neighbourhood of the tissue.  Crucially, cells of
\emph{all} types within the same bin are considered to share a micro-environment
and can thus interact.
}
\keyword{internal}
