% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_pairs.R
\name{.build_result}
\alias{.build_result}
\title{Build standardised scPairs result object}
\usage{
.build_result(pair_dt, features, object, has_spatial, params)
}
\description{
Build standardised scPairs result object
}
\keyword{internal}
