% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_celltype.R
\name{.cross_subtitle}
\alias{.cross_subtitle}
\title{Format subtitle with global cross-cell-type stats}
\usage{
.cross_subtitle(score, r_ab, r_ba)
}
\description{
Format subtitle with global cross-cell-type stats
}
\keyword{internal}
