% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{.extract_pair_vals}
\alias{.extract_pair_vals}
\title{Extract pairwise values from a symmetric matrix given a pair table}
\usage{
.extract_pair_vals(sym_mat, pair_dt)
}
\description{
Extract pairwise values from a symmetric matrix given a pair table
}
\keyword{internal}
