% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_knowledge.R
\name{.get_go_gene_sets}
\alias{.get_go_gene_sets}
\title{Retrieve GO Biological Process gene sets}
\usage{
.get_go_gene_sets(organism, genes = NULL, min_genes = 5, max_genes = 500)
}
\description{
Retrieve GO Biological Process gene sets
}
\keyword{internal}
