% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_knowledge.R
\name{.get_kegg_gene_sets}
\alias{.get_kegg_gene_sets}
\title{Retrieve KEGG pathway gene sets}
\usage{
.get_kegg_gene_sets(organism, genes = NULL, min_genes = 5, max_genes = 500)
}
\description{
Retrieve KEGG pathway gene sets
}
\keyword{internal}
