% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_spatial}
\alias{.has_spatial}
\title{Detect whether a Seurat object contains spatial information}
\usage{
.has_spatial(object)
}
\arguments{
\item{object}{A Seurat object.}
}
\value{
Logical scalar.
}
\description{
Detect whether a Seurat object contains spatial information
}
\keyword{internal}
