% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.msg}
\alias{.msg}
\title{Print a progress message (respects verbose flag)}
\usage{
.msg(..., verbose = TRUE)
}
\arguments{
\item{...}{Parts of the message (passed to \code{paste0}).}

\item{verbose}{Logical.}
}
\description{
Print a progress message (respects verbose flag)
}
\keyword{internal}
