% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{.neighbourhood_coexpr}
\alias{.neighbourhood_coexpr}
\title{Single-pair neighbourhood co-expression score}
\usage{
.neighbourhood_coexpr(x, y, W, k = NULL)
}
\description{
Single-pair neighbourhood co-expression score
}
\keyword{internal}
