% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_knowledge.R
\name{.neighbourhood_synergy}
\alias{.neighbourhood_synergy}
\title{Single-pair neighbourhood synergy score}
\usage{
.neighbourhood_synergy(x, y, W)
}
\description{
Single-pair neighbourhood synergy score
}
\keyword{internal}
