% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_knowledge.R
\name{.prior_score}
\alias{.prior_score}
\title{Compute prior interaction score for a single gene pair}
\usage{
.prior_score(gene1, gene2, prior_net)
}
\description{
Compute prior interaction score for a single gene pair
}
\keyword{internal}
