% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_pairs.R
\name{.resolve_cluster_ids}
\alias{.resolve_cluster_ids}
\title{Resolve cluster IDs from Seurat object}
\usage{
.resolve_cluster_ids(object, cluster_col = NULL)
}
\description{
Resolve cluster IDs from Seurat object
}
\keyword{internal}
