% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{.smoothed_cor}
\alias{.smoothed_cor}
\title{Compute smoothed correlation for a single gene pair}
\usage{
.smoothed_cor(x, y, W, alpha = 0.3)
}
\description{
Compute smoothed correlation for a single gene pair
}
\keyword{internal}
