% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.validate_min_cells_expressed}
\alias{.validate_min_cells_expressed}
\title{Validate minimum cells expressed parameter}
\usage{
.validate_min_cells_expressed(min_cells_expressed, n_cells)
}
\description{
Validate minimum cells expressed parameter
}
\keyword{internal}
