% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.validate_neighbourhood_params}
\alias{.validate_neighbourhood_params}
\title{Validate neighbourhood parameters}
\usage{
.validate_neighbourhood_params(neighbourhood_k, n_cells)
}
\description{
Validate neighbourhood parameters
}
\keyword{internal}
