% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_seurat}
\alias{.validate_seurat}
\title{Validate that input is a Seurat object}
\usage{
.validate_seurat(object, require_spatial = FALSE)
}
\arguments{
\item{object}{Object to validate.}

\item{require_spatial}{Logical; if TRUE, require spatial coordinates.}
}
\value{
TRUE invisibly; stops with informative error otherwise.
}
\description{
Validate that input is a Seurat object
}
\keyword{internal}
