% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null-coalescing operator (if not already imported from Seurat)}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value.}

\item{b}{Second value (returned if \code{a} is NULL).}
}
\value{
\code{a} if not NULL, otherwise \code{b}.
}
\description{
Null-coalescing operator (if not already imported from Seurat)
}
\keyword{internal}
