% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pair.R
\name{print.scPairs_pair_result}
\alias{print.scPairs_pair_result}
\title{Print method for scPairs_pair_result}
\usage{
\method{print}{scPairs_pair_result}(x, ...)
}
\arguments{
\item{x}{An scPairs_pair_result object.}

\item{...}{Ignored.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for scPairs_pair_result
}
