% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_pairs.R
\name{print.scPairs_result}
\alias{print.scPairs_result}
\title{Print method for scPairs_result}
\usage{
\method{print}{scPairs_result}(x, ...)
}
\arguments{
\item{x}{An scPairs_result object.}

\item{...}{Ignored.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for scPairs_result
}
