% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{RAW_METRICS}
\alias{RAW_METRICS}
\title{Metrics used as raw values (not absolute-valued)}
\format{
An object of class \code{character} of length 5.
}
\usage{
RAW_METRICS
}
\description{
Metrics used as raw values (not absolute-valued)
}
\keyword{internal}
