% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{RESULT_COLUMNS}
\alias{RESULT_COLUMNS}
\title{Standard column names for result data.tables}
\format{
An object of class \code{list} of length 6.
}
\usage{
RESULT_COLUMNS
}
\description{
Standard column names for result data.tables
}
\keyword{internal}
