% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{SCORE_CONFIDENCE_QUANTILES}
\alias{SCORE_CONFIDENCE_QUANTILES}
\title{Score-based confidence thresholds (when no p-values available)}
\format{
An object of class \code{list} of length 3.
}
\usage{
SCORE_CONFIDENCE_QUANTILES
}
\description{
Score-based confidence thresholds (when no p-values available)
}
\keyword{internal}
